% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_identify.R
\name{check_identification}
\alias{check_identification}
\title{In progress: check for identifiability of stochastic interventions}
\usage{
check_identification(
  X,
  delta = 0.1,
  Xdensity_learners = NULL,
  Xbinary_learners = NULL,
  verbose = FALSE,
  scale_continuous = TRUE,
  ...
)
}
\arguments{
\item{X}{data frame of predictors}

\item{delta}{change in each column of X corresponding to stochastic intervention}

\item{Xdensity_learners}{list of sl3 learners used to estimate the density of continuous predictors, conditional on all other predictors in X}

\item{Xbinary_learners}{list of sl3 learners used to estimate the probability mass of continuous predictors, conditional on all other predictors in X}

\item{verbose}{(logical) print extra information}

\item{...}{passed to sl3::make_sl3_Task (e.g. weights)}
}
\description{
In progress: check for identifiability of stochastic interventions
}
\examples{
\dontrun{
data(metals, package="qgcomp")
XYlist = list(X=metals[,1:23], Y=metals$y)
Xbinary_learners = .default_binary_learners()
Xdensity_learners = .default_density_learners(n_bins=c(5, 20))
ident <- check_identification(X=XYlist$X, delta=0.1, Y_learners = Y_learners,
       Xdensity_learners=Xdensity_learners[1:2], Xbinary_learners=Xbinary_learners,
       estimator="TMLE", verbose=TRUE)
ident
}



}
