% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{.dgm}
\alias{.dgm}
\title{Title}
\usage{
.dgm(n, p, ncat)
}
\arguments{
\item{n}{sample size}

\item{p}{number of predictors (with random correlation matrix)}

\item{ncat}{number of dichotomous predictors (must be <= p)}
}
\value{

}
\description{
Title
}
