% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_utility.R
\name{.train_superlearner}
\alias{.train_superlearner}
\title{.train_superlearner}
\usage{
.train_superlearner(
  datatask,
  learners,
  type = c("density", "probability", "expectation")
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Train a super learner fit, given some data held in an sl3 task object and a list of learners
}
