% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_aipw.R
\name{.varimp_aipw}
\alias{.varimp_aipw}
\title{Title}
\usage{
.varimp_aipw(
  X,
  Y,
  delta = 0.1,
  Y_learners = NULL,
  Xdensity_learners = NULL,
  Xbinary_learners = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
XYlist = .dgm(n=100,p=4,ncat=3)
data(metals, package="qgcomp")
XYlist = list(X=metals[,1:23], Y=metals$y)
Y_learners = .default_continuous_learners()
Xbinary_learners = .default_binary_learners()
Xdensity_learners = .default_density_learners(n_bins=c(5, 20))
vi <- .varimp_aipw(X=XYlist$X,Y=XYlist$Y, delta=0.1, Y_learners = Y_learners[1:4],Xdensity_learners=Xdensity_learners[1:2], Xbinary_learners=Xbinary_learners[1:2] )
vi
}
}
