% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_gcomp.R
\name{.varimp_gcomp}
\alias{.varimp_gcomp}
\title{Variable importance using g-computation}
\usage{
.varimp_gcomp(
  X,
  Y,
  delta = 0.1,
  Y_learners = NULL,
  Xdensity_learners = NULL,
  Xbinary_learners = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{X}{data frame of predictors}

\item{Y}{outcome}

\item{delta}{change in each column of X corresponding to}

\item{Y_learners}{list of sl3 learners used to predict the outcome, conditional on all predictors in X}

\item{Xdensity_learners}{list of sl3 learners used to estimate the density of continuous predictors, conditional on all other predictors in X}

\item{Xbinary_learners}{list of sl3 learners used to estimate the probability mass of continuous predictors, conditional on all other predictors in X}

\item{verbose}{(logical) print extra information}

\item{...}{passed to sl3::base_predict (rare)}
}
\value{
vi object
}
\description{
Not usually called by users
}
