% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_tmle.R
\name{.varimp_tmle}
\alias{.varimp_tmle}
\title{Variable importance using targeted minimum loss estimation}
\usage{
.varimp_tmle(
  X,
  Y,
  delta = 0.1,
  Y_learners = NULL,
  Xdensity_learners = NULL,
  Xbinary_learners = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{X}{data frame of predictors}

\item{Y}{outcome}

\item{delta}{change in each column of X corresponding to}

\item{Y_learners}{list of sl3 learners used to predict the outcome, conditional on all predictors in X}

\item{Xdensity_learners}{list of sl3 learners used to estimate the density of continuous predictors, conditional on all other predictors in X}

\item{Xbinary_learners}{list of sl3 learners used to estimate the probability mass of continuous predictors, conditional on all other predictors in X}

\item{verbose}{(logical) print extra information}

\item{...}{passed to sl3::base_predict (rare)}
}
\value{
vi object
}
\description{
Not usually called by users
}
\examples{
\dontrun{
XYlist = .dgm(n=100,p=4,ncat=3)
data(metals, package="qgcomp")
XYlist = list(X=metals[,1:23], Y=metals$y)
Y_learners = .default_continuous_learners()
Xbinary_learners = .default_binary_learners()
Xdensity_learners = .default_density_learners(n_bins=c(5, 20))
vi <- .varimp_aipw(X=XYlist$X,Y=XYlist$Y, delta=0.1, Y_learners = Y_learners[1:4],
Xdensity_learners=Xdensity_learners[1:2], Xbinary_learners=Xbinary_learners[1:2] )
vi
}
}
