% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_identify.R
\name{plotshift_dens}
\alias{plotshift_dens}
\title{In progress: visualizing a stochastic intervention}
\usage{
plotshift_dens(vibr.fit, X, Acol = 1, delta)
}
\arguments{
\item{vibr.fit}{a fit from varimp}

\item{X}{predictors from a varimp fit}

\item{Acol}{(integer) which column of X to diagnose}

\item{delta}{(numeric, default=0.01) change in each column of X corresponding to stochastic intervention}
}
\value{
ggplot2 plot object
}
\description{
In progress: visualizing a stochastic intervention
}
\examples{
set.seed(123)
}
