% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{varimp}
\alias{varimp}
\title{Variable importance in cross-sectional data}
\usage{
varimp(
  X,
  Y,
  delta = 0.1,
  Y_learners = NULL,
  Xdensity_learners = NULL,
  Xbinary_learners = NULL,
  verbose = TRUE,
  estimator = "AIPW",
  bounded = FALSE,
  estimand = "diff",
  B = NULL,
  showProgress = TRUE,
  scale_continuous = TRUE,
  ...
)
}
\arguments{
\item{X}{data frame of predictors}

\item{Y}{outcome}

\item{delta}{change in each column of X corresponding to}

\item{Y_learners}{list of sl3 learners used to predict the outcome, conditional on all predictors in X}

\item{Xdensity_learners}{list of sl3 learners used to estimate the density of continuous predictors, conditional on all other predictors in X}

\item{Xbinary_learners}{list of sl3 learners used to estimate the probability mass of continuous predictors, conditional on all other predictors in X}

\item{verbose}{(logical) print extra information}

\item{estimator}{(character) "AIPW" (default), "GCOMP", "IPW"}

\item{estimand}{(character) "diff" (default, estimate mean difference comparing Y under intervention with observed Y), "mean" (estimate mean Y under intervention)}

\item{B}{(NULL or integer) Numer of bootstrap iterations (NULL = asymptotic variance only)}

\item{showProgress}{show progress of bootstrapping (only relevant if B is not NULL)}

\item{...}{passed to sl3::make_sl3_Task (e.g. weights)}
}
\value{
vi object
}
\description{
Variable importance in cross-sectional data
}
\examples{
\dontrun{
data(metals, package="qgcomp")
XYlist = list(X=metals[,1:23], Y=metals$y)
Y_learners = .default_continuous_learners_big()
Xbinary_learners = .default_binary_learners()
Xdensity_learners = .default_density_learners(n_bins=c(5, 20))
vi <- varimp(X=XYlist$X,Y=XYlist$Y, delta=0.1, Y_learners = Y_learners,
       Xdensity_learners=Xdensity_learners[1:2], Xbinary_learners=Xbinary_learners,
       estimator="TMLE")
vi
}
}
