% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_identify.R
\name{plotshift_scatter}
\alias{plotshift_scatter}
\title{Title}
\usage{
plotshift_scatter(vibr.fit, Acol, Bcol, delta = NULL, joint = FALSE)
}
\arguments{
\item{delta}{}
}
\value{

}
\description{
Title
}
