% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{varimp}
\alias{varimp}
\title{Variable importance in cross-sectional data}
\usage{
varimp(
  X,
  Y,
  delta = 0.1,
  Y_learners = NULL,
  Xdensity_learners = NULL,
  Xbinary_learners = NULL,
  verbose = TRUE,
  estimator = "AIPW",
  ...
)
}
\arguments{
\item{X}{data frame of predictors}

\item{Y}{outcome}

\item{delta}{change in each column of X corresponding to}

\item{Y_learners}{list of sl3 learners used to predict the outcome, conditional on all predictors in X}

\item{Xdensity_learners}{list of sl3 learners used to estimate the density of continuous predictors, conditional on all other predictors in X}

\item{Xbinary_learners}{list of sl3 learners used to estimate the probability mass of continuous predictors, conditional on all other predictors in X}

\item{verbose}{(logical) print extra information}

\item{estimator}{(character) "AIPW" (default), "GCOMP", "IPW"}

\item{...}{passed to sl3::make_sl3_Task (e.g. weights)}
}
\value{
vi object
}
\description{
Variable importance in cross-sectional data
}
