% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{varimp}
\alias{varimp}
\title{Variable importance in cross-sectional data}
\usage{
varimp(
  X,
  W = NULL,
  Y,
  V = NULL,
  delta = 0.1,
  Y_learners = NULL,
  Xdensity_learners = NULL,
  Xbinary_learners = NULL,
  verbose = TRUE,
  estimator = "AIPW",
  bounded = FALSE,
  updatetype = "weighted",
  estimand = "diff",
  family = gaussian(),
  xfitfolds = 3,
  foldrepeats = 10,
  B = NULL,
  showProgress = TRUE,
  scale_continuous = TRUE,
  ...
)
}
\arguments{
\item{X}{data frame of variables for which variable importance will be estimated}

\item{W}{data frame of covariates (e.g. potential confounders) for which variable importance will not be estimated}

\item{Y}{outcome}

\item{V}{(default NULL) a data frame of other variables that contain weights or offsets, if used ("weights" and "offset" are both passed to other functions via extra arguments represented by ...: see example below)}

\item{delta}{change in each column of X corresponding to}

\item{Y_learners}{list of sl3 learners used to predict the outcome, conditional on all predictors in X}

\item{Xdensity_learners}{list of sl3 learners used to estimate the density of continuous predictors, conditional on all other predictors in X}

\item{Xbinary_learners}{list of sl3 learners used to estimate the probability mass of continuous predictors, conditional on all other predictors in X}

\item{verbose}{(logical) print extra information}

\item{estimator}{(character) "AIPW" (default), "TMLE", "GCOMP", "IPW", "TMLEX" (cross fit TMLE), "AIPWX" (cross fit AIPW)}

\item{bounded}{(logical) not used}

\item{updatetype}{(character) (used for estimator = "TMLE" only) "weighted" or any other valid character. If "weighted" then uses weighting by clever covariate in update step of TMLE, otherwise fits a generalized linear model with no intercept and clever covariate as a sole predictor}

\item{estimand}{(character) "diff" (default, estimate mean difference comparing Y under intervention with observed Y), "mean" (estimate mean Y under intervention)}

\item{family}{(character or glm families binomial or gaussian, default = gaussian()) Outcome type: can be gaussian(), binomial(), "gaussian" or "binomial"; will be guessed if left NULL}

\item{xfitfolds}{(odd integer, default=3) (used for estimator = "TMLEX" only) number of cross-fit folds (must be odd number - last fold is used for validation while the rest of the data are split in two for fitting treatment or outcome models)}

\item{foldrepeats}{(integer, default=10) (used for estimator = "TMLEX" only) number of times to repeat cross-fitting (higher numbers = more stable)}

\item{B}{(NULL or integer) Number of bootstrap iterations (NULL = asymptotic variance only)}

\item{showProgress}{show progress of bootstrapping (only relevant if B is not NULL)}

\item{...}{passed to sl3::make_sl3_Task (e.g. weights)}
}
\value{
vi object
}
\description{
Variable importance in cross-sectional data
}
\examples{
\dontrun{
data(metals, package="qgcomp")
XYlist = list(X=metals[,1:23], Y=metals$y)
Y_learners = .default_continuous_learners()
Xbinary_learners = list(Lrnr_stepwise$new(name="SW"))
Xdensity_learners = .default_density_learners(n_bins=c(10))
vi <- varimp(X=XYlist$X,Y=XYlist$Y, delta=0.1, Y_learners = Y_learners,
       Xdensity_learners=Xdensity_learners[1:2], Xbinary_learners=Xbinary_learners,
       estimator="TMLE")
vi
V = data.frame(wt=runif(nrow(metals)))
viw <- varimp(X=XYlist$X,Y=XYlist$Y, V=V, delta=0.1, Y_learners = Y_learners,
       Xdensity_learners=Xdensity_learners[1:2], Xbinary_learners=Xbinary_learners,
       estimator="TMLE", weights="wt")
viw
}
}
